Single activity impl

+++ app/src/main/java/com/example/daggerhilttest/remote/WeatherApi.kt
@@ -1,6 +1,5 @@
 package com.example.daggerhilttest.remote
 
-import com.example.daggerhilttest.constants.Constants
 import com.example.daggerhilttest.models.v1.CurrentWeather
 import com.example.daggerhilttest.models.v1.HourlyForecast
 import com.example.daggerhilttest.models.v2.GeocodingResponseV2
@@ -15,8 +14,8 @@ interface WeatherApi {
 
     @GET("VisualCrossingWebServices/rest/services/timeline/{latitude},{longitude}")
     suspend fun getWeatherData(
-        @Path("latitude") latitude: Double,
-        @Path("longitude") longitude: Double,
+        @Path("latitude") latitude: Float,
+        @Path("longitude") longitude: Float,
         @Query("key") apiKey: String,
     ): Response<WeatherResponseV2>
 
@@ -24,8 +23,8 @@ interface WeatherApi {
     @GET
     suspend fun getLocationFromGeocoding(
         @Url url: String,
-        @Query("lat") lat: Double,
-        @Query("lon") long: Double,
+        @Query("lat") lat: Float,
+        @Query("lon") long: Float,
         @Query("appid") apiKey: String,
     ): Response<List<GeocodingResponseV2>>
